Grouped and moved functions into separate files (#10)
- moved all logging functions from main.rs to log.rs
	- only moved init_log and all the use statements
- moved all user validation functions, and json save and load function to users.rs
